requirements.txt
fastapi
starlette
chromadb
pysqlite3-binary
langchain
langchain-core
langchain-community
langgraph
lark
tiktoken
streamlit
pypdf
ollama
sentence-transformers